Allow rendering mimir templates in other contexts
When Mimir is used as a subchart, the rendering of the annotation
"checksum/config" on PODs is wrong. This is important to trigger
service restart when the config is changed.

There are two options: config is in Mimir or shared with mimir in which
case it can contain templates and values of the parent chart, making it
impossible to evaluate in the subchart. Or the config lives in the parent
chart, but then it is not possible to reach it in Mimir as subchart and
the checksum is checksum of empty string. See: "few important details" in
https://helm.sh/docs/chart_template_guide/subcharts_and_globals/

This patch makes the manifest templates global named templates and thus
able to be rendered in the parent context. It is one step short of
merging the mimir-distributed chart with enterprise-metrics.

Signed-off-by: György Krajcsovits <gyorgy.krajcsovits@grafana.com>